# target/swagger.json
{"swagger":"2.0","info":{"version":"1.0.0","title":"ONOS SSH Client REST API","description":"APIs for interacting with the SSH Client."},"basePath":"/onos/v1/sshrest","produces":["application/json"],"consumes":["application/json"],"tags":[{"name":"devices"}],"paths":{"/devices":{"get":{"tags":["devices"],"parameters":[],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/groups/{switchName}":{"get":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/scommand/{serverName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"serverName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/execSudoCommand"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/vxlanloopback/{switchName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"switchName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/setVxlanLoopback"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/command/{serverName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"serverName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/execCommand"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/log/{switchName}":{"get":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true},{"name":"filename","in":"query","type":"string","required":true}],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/vxlanvlan/{switchName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"switchName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/setVxlanVlan"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/flows/{switchName}":{"get":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/controller/{switchName}":{"get":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true}],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"switchName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/setController"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"delete":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"switchName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation","schema":{"$ref":"#/definitions/unsetController"}},"default":{"description":"Unexpected error"}}}},"/devices/ssid/{apName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"apName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/setSSID"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/vtep/{switchName}":{"post":{"tags":["devices"],"summary":"","description":"","parameters":[{"name":"switchName","in":"path","type":"string","description":"","required":true},{"name":"stream","in":"body","schema":{"$ref":"#/definitions/setVtep"},"description":"","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}},"/devices/vxlan/{switchName}":{"post":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true},{"name":"status","in":"query","type":"boolean","required":true},{"name":"stream","in":"body","required":true}],"consumes":["application/json"],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}},"get":{"tags":["devices"],"parameters":[{"name":"switchName","in":"path","type":"string","required":true}],"produces":["application/json"],"responses":{"200":{"description":"successful operation"},"default":{"description":"Unexpected error"}}}}},"definitions":{"setController":{"type":"object","required":["ip"],"properties":{"ip":{"type":"string","description":"controller IPv4 address","example":"192.168.20.76"},"port":{"type":"string","description":"controller port","example":"6653"}}},"unsetController":{"type":"object","required":["ip"],"properties":{"ip":{"type":"string","description":"controller IPv4 address","example":"192.168.20.76"}}},"execCommand":{"type":"object","required":["command"],"properties":{"command":{"type":"string","description":"non-privileged command","example":"echo 123"}}},"execSudoCommand":{"type":"object","required":["scommand"],"properties":{"scommand":{"type":"string","description":"privileged command","example":"ovs-vsctl show"}}},"setSSID":{"type":"object","required":["ssid"],"properties":{"ssid":{"type":"string","description":"AP SSID","example":"dlink-1"}}},"setVxlanLoopback":{"type":"object","required":["id"],"properties":{"id":{"type":"string","description":"source-loopback ID in switch for VXLAN","example":"0"}}},"setVxlanVlan":{"type":"object","required":["vnid","vid"],"properties":{"vnid":{"type":"string","description":"VNI","example":"2000"},"vid":{"type":"string","description":"VLAN ID","example":"10"}}},"setVtep":{"type":"object","required":["vnid","ip"],"properties":{"vnid":{"type":"string","description":"VXLAN Network ID","example":"2000"},"ip":{"type":"string","description":"VTEP IP address","example":"10.0.0.1"},"mac":{"type":"string","description":"VTEP MAC address","example":"ab:cd:ef:01:23:45"}}}}}